requirements.txt
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
cachetools==4.2.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
idna==3.3
kombu==5.2.0
Pillow==8.3.2
prompt-toolkit==3.0.21
psycopg2==2.9.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-telegram-bot==13.8.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
